feat: add deployment analytics endpoint and integrate deployment routes
